requirements.txt
openai
python-dotenv
gradio
PyPDF2
streamlit
azure-kusto-data
pyautogen
vision-agent
azure-cognitiveservices-speech
bs4
streamlit-audiorecorder
pyaudio
azure-ai-translation-text==1.0.0b1
PyQt5
PyMuPDF
python-docx
matplotlib
Pillow
streamlit_quill
semantic-kernel
azure-search-documents